Redesigned Assignment Screen and added skeleton loading
