+++ app/src/main/java/com/disfluency/screens/phono/AssingmentScreen.kt
@@ -11,44 +11,48 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
+import com.disfluency.components.list.items.ExerciseListItem
+import com.disfluency.components.list.items.PatientListItem
 import com.disfluency.data.ExerciseRepository
 import com.disfluency.data.PatientRepository
+import com.disfluency.loading.SkeletonLoader
+import com.disfluency.loading.skeleton.exercise.ExerciseListItemSkeleton
+import com.disfluency.loading.skeleton.patient.PatientListItemSkeleton
 import com.disfluency.model.Exercise
 import com.disfluency.model.Patient
 import com.disfluency.model.Phono
 import com.disfluency.model.dto.AssignmentDTO
 import com.disfluency.navigation.Route
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
+import kotlinx.coroutines.*
 
 @Composable
 fun AssignmentScreen(navController: NavHostController, phono: Phono) {
     val checkedPatients = remember { mutableStateListOf<String>() }
     val checkedExercises = remember { mutableStateListOf<String>() }
 
-    val patients = remember { mutableStateListOf<Patient>() }
-    val exercises = remember { mutableStateListOf<Exercise>() }
+    val patients: MutableState<List<Patient>?> = remember { mutableStateOf(null) }
+    val exercises: MutableState<List<Exercise>?> = remember { mutableStateOf(null) }
 
     LaunchedEffect(Unit) {
         val patientsResponse = withContext(Dispatchers.IO) { PatientRepository.getPatientsByTherapistId(phono.id) }
         Log.i("HTTP", patientsResponse.toString())
-        patients.addAll(patientsResponse)
+        patients.value = patientsResponse
     }
 
     LaunchedEffect(Unit) {
         val exercisesResponse = withContext(Dispatchers.IO) { ExerciseRepository.getExercisesByTherapistId(phono.id) }
         Log.i("HTTP", exercisesResponse.toString())
-        exercises.addAll(exercisesResponse)
+        exercises.value = exercisesResponse
     }
 
     Column(modifier = Modifier.fillMaxSize()) {
         Text(
             text = "Pacientes",
-            style = MaterialTheme.typography.headlineSmall,
+            style = MaterialTheme.typography.titleLarge,
+            fontWeight = FontWeight.SemiBold,
             modifier = Modifier
                 .padding(16.dp)
         )
@@ -60,12 +64,13 @@ fun AssignmentScreen(navController: NavHostController, phono: Phono) {
                     .padding(8.dp),
                 contentAlignment = Alignment.Center
             ) {
-                PatientsColumn(patients, checkedPatients, navController)
+                PatientsColumn(patients, checkedPatients)
             }
         }
         Text(
             text = "Ejercicios",
-            style = MaterialTheme.typography.headlineSmall,
+            style = MaterialTheme.typography.titleLarge,
+            fontWeight = FontWeight.SemiBold,
             modifier = Modifier
                 .padding(16.dp)
         )
@@ -77,7 +82,7 @@ fun AssignmentScreen(navController: NavHostController, phono: Phono) {
                     .padding(8.dp),
                 contentAlignment = Alignment.Center
             ) {
-                ExercisesColumn(exercises, checkedExercises, navController)
+                ExercisesColumn(exercises, checkedExercises)
             }
         }
     }
@@ -90,79 +95,104 @@ fun AssignmentScreen(navController: NavHostController, phono: Phono) {
 }
 
 @Composable
-fun PatientsColumn(patients: List<Patient>, checkedPatients: MutableList<String>, navController: NavHostController) {
+fun PatientsColumn(patients: MutableState<List<Patient>?>, checkedPatients: MutableList<String>) {
     Column(modifier = Modifier.fillMaxHeight()) {
         Row(modifier = Modifier.fillMaxWidth()) {
-            LazyColumn(
-                verticalArrangement = Arrangement.spacedBy(4.dp)){
-                items(patients){ patient ->
-                    PatientAssignmentCard(patient, checkedPatients, navController)
+
+            SkeletonLoader(
+                state = patients,
+                content = {
+                    patients.value?.let {
+                        LazyColumn(verticalArrangement = Arrangement.spacedBy(4.dp)){
+                            items(it){ patient ->
+                                PatientAssignmentListItem(patient, checkedPatients)
+                            }
+                        }
+                    }
+                },
+                skeleton = {
+                    LazyColumn(verticalArrangement = Arrangement.spacedBy(4.dp)){
+                        items(3){
+                            PatientListItemSkeleton()
+                        }
+                    }
                 }
-            }
+            )
+
         }
     }
 }
 
 @Composable
-fun ExercisesColumn(exercises: List<Exercise>, checkedExercises: MutableList<String>, navController: NavHostController) {
+fun ExercisesColumn(exercises: MutableState<List<Exercise>?>, checkedExercises: MutableList<String>) {
     Column(modifier = Modifier.fillMaxHeight()) {
         Row(modifier = Modifier.fillMaxWidth()) {
-            LazyColumn(
-                verticalArrangement = Arrangement.spacedBy(4.dp)){
-                items(exercises){ ex ->
-                    ExerciseAssignmentCard(ex, checkedExercises, navController)
+
+            SkeletonLoader(
+                state = exercises,
+                content = {
+                    exercises.value?.let {
+                        LazyColumn(verticalArrangement = Arrangement.spacedBy(4.dp)){
+                            items(it){ ex ->
+                                ExerciseAssignmentListItem(ex, checkedExercises)
+                            }
+                        }
+                    }
+                },
+                skeleton = {
+                    LazyColumn(verticalArrangement = Arrangement.spacedBy(4.dp)){
+                        items(3){
+                            ExerciseListItemSkeleton()
+                        }
+                    }
                 }
-            }
+            )
         }
     }
 }
 
 @Composable
-fun PatientAssignmentCard(patient: Patient, checkedPatients: MutableList<String>, navController: NavHostController) {
-    Row(modifier = Modifier.padding(2.dp)) {
-        val isChecked = remember { mutableStateOf(patient.id in checkedPatients) }
-        Checkbox(
-            checked = isChecked.value,
-            onCheckedChange =  {
-                isChecked.value = it
-                if (isChecked.value) {
-                    if (!checkedPatients.contains(patient.id))
-                        checkedPatients.add(patient.id)
-                } else {
-                    checkedPatients.remove(patient.id)
+fun PatientAssignmentListItem(patient: Patient, checkedPatients: MutableList<String>) {
+    val isChecked = remember { mutableStateOf(patient.id in checkedPatients) }
+
+    PatientListItem(patient = patient, leadingContentPrefix = {
+            Checkbox(
+                modifier = Modifier.offset((-8).dp, 0.dp),
+                checked = isChecked.value,
+                onCheckedChange =  {
+                    isChecked.value = it
+                    if (isChecked.value) {
+                        if (!checkedPatients.contains(patient.id))
+                            checkedPatients.add(patient.id)
+                    } else {
+                        checkedPatients.remove(patient.id)
+                    }
                 }
-            })
-        ListItem(
-            headlineContent = { Text(text = patient.fullName()) },
-            modifier = Modifier.clickable {
-                navController.navigate(Route.Paciente.routeTo(patient.id))
-            }
-        )
-    }
+            )
+        }
+    )
 }
 
 @Composable
-fun ExerciseAssignmentCard(exercise: Exercise, checkedExercises: MutableList<String>, navController: NavHostController) {
-    Row(modifier = Modifier.padding(2.dp)) {
-        val isChecked = remember { mutableStateOf(exercise.id in checkedExercises) }
-        Checkbox(
-            checked = isChecked.value,
-            onCheckedChange =  {
-                isChecked.value = it
-                if (isChecked.value) {
-                    if (!checkedExercises.contains(exercise.id))
-                        checkedExercises.add(exercise.id)
-                } else {
-                    checkedExercises.remove(exercise.id)
+fun ExerciseAssignmentListItem(exercise: Exercise, checkedExercises: MutableList<String>) {
+    val isChecked = remember { mutableStateOf(exercise.id in checkedExercises) }
+
+    ExerciseListItem(exercise = exercise, leadingContentPrefix = {
+            Checkbox(
+                modifier = Modifier.offset((-8).dp, 0.dp),
+                checked = isChecked.value,
+                onCheckedChange =  {
+                    isChecked.value = it
+                    if (isChecked.value) {
+                        if (!checkedExercises.contains(exercise.id))
+                            checkedExercises.add(exercise.id)
+                    } else {
+                        checkedExercises.remove(exercise.id)
+                    }
                 }
-            })
-        ListItem(
-            headlineContent = { Text(text = exercise.title) },
-            modifier = Modifier.clickable {
-                navController.navigate(Route.Ejercicio.routeTo(exercise.id))
-            }
-        )
-    }
+            )
+        }
+    )
 }
 
 @Composable
@@ -186,6 +216,7 @@ fun AssignButton(
                 }
             },
             modifier = Modifier.padding(16.dp),
+            containerColor = MaterialTheme.colorScheme.secondary
         ) {
             Icon(Icons.Filled.Send, "Asignar")
         }
