+++ app/src/main/java/com/disfluency/screens/phono/PatientsScreen.kt
@@ -27,6 +27,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
 import androidx.navigation.NavHostController
 import com.disfluency.R
+import com.disfluency.components.list.items.PatientListItem
 import com.disfluency.components.user.IconLabeled
 import com.disfluency.components.user.weeklyTurnFormat
 import com.disfluency.data.PatientRepository
@@ -100,59 +101,14 @@ fun PatientsList(patients: List<Patient>, navController: NavHostController, filt
         verticalArrangement = Arrangement.spacedBy(8.dp)) {
         items(patients.filter {
                 patient -> patient.fullName().contains(filter, true) }) {patient ->
-            PatientCard(patient, navController)
+            PatientListItem(patient, onClick = {
+                navController.navigate(Route.Paciente.routeTo(patient.id))
+            })
         }
     }
 }
 
-@Composable
-fun PatientCard(patient: Patient, navController: NavHostController) {
-    val onClick = {
-        navController.navigate(Route.Paciente.routeTo(patient.id))
-    }
 
-    Card(
-        modifier = Modifier.clickable { onClick() },
-        border = BorderStroke(1.dp, MaterialTheme.colorScheme.primary)
-    ) {
-        ListItem(
-            modifier = Modifier.height(56.dp),
-            headlineContent = {
-                Text(
-                    text = patient.fullNameFormal(),
-                    style = MaterialTheme.typography.titleMedium
-                )
-            },
-            supportingContent = {
-                Text(
-                    text = weeklyTurnFormat(patient.weeklyTurn),
-                    style = MaterialTheme.typography.labelMedium
-                )
-            },
-            leadingContent = {
-                Image(
-                    painter = painterResource(patient.profilePic),
-                    contentDescription = null,
-                    modifier = Modifier
-                        .size(40.dp)
-                        .clip(CircleShape)
-                        .border(1.5.dp, MaterialTheme.colorScheme.secondary, CircleShape)
-                )
-            },
-            trailingContent = {
-                IconLabeled(
-                    icon = Icons.Outlined.AccessTime,
-                    label = patient.weeklyHour.format(
-                        DateTimeFormatter.ofPattern(stringResource(
-                        R.string.time_format))),
-                    content = "Time"
-                )
-            }
-        )
-    }
-
-
-}
 
 @Composable
 fun PatientCreation(navController: NavHostController) {
